Проведен успешно рефакторинг: создан новый модуль tools, в который вынесены самые используемые функции и запросы.
