+++ ai_model/templates/show_quality.html
@@ -16,9 +16,9 @@
         </div>
         <div class="section">
             <ul>
-                <li><h2>Значение метрики по температуре</h2></li>
+                <h2>Значение метрики по температуре</h2>
                 <li><p><strong>WAPE</strong> = {{ metrics[0] }}</p></li>
-                <li><h2>Значения метрик по заболеваемости</h2></li>
+                <h2>Значения метрик по заболеваемости</h2>
                 <li><p><strong>RMSE</strong> = {{ metrics[1] }}</p></li>
                 <li><p><strong>MAE</strong> = {{ metrics[2] }}</p></li>
             </ul>
